Move TextprotoFilePrincipalLookup.kt to measurement/api/v2alpha.
